add partner section and refactor about section to remove partners prop
