+++ app/page.tsx
@@ -12,6 +12,7 @@ import { ContactSectionWrapper } from "@/components/sections/contact-section";
 import { FooterSection } from "@/components/sections/footer-section";
 import { NavbarComponent } from "@/components/sections/nav-bar";
 import { AuroraBackground } from "@/components/ui/aurora-background";
+import PartnerSection from "@/components/sections/partner-section";
 
 export default function Home() {
   const [activeSection, setActiveSection] = useState("home");
@@ -145,10 +146,11 @@ export default function Home() {
         {/* Content wrapper with higher z-index */}
         <div className="relative" style={{ zIndex: 30 }}>
           <HeroSection />
-          <AboutSection partners={partners} />
+          <AboutSection />
           <TimelineSection events={timelineEvents} />
           <GallerySection images={galleryImages} />
           <CompetitionSection />
+          <PartnerSection partners={partners} />
           <TeamSection teamMembers={teamMembers} />
           <ContactSectionWrapper teamMembers={teamMembers} />
           <FooterSection />
